Add some comments
In this commit I just tried to understand the flow of the code and added
documentation and comments to help guide my way. These are required and are part
of the grade, so more internal comments will be required to sign post the TAs
and graders.

Future commits will hopefully clean up the code a little, to make it easier to
follow.