Major overhaul of how we do allocation.
The FreeBlockList now does the Best Fit algorithm internally, so there is no
longer an external MMAlg object that acts on the FBL. This increases coupling,
but both makes the code simpler and faster. It also eliminates the code smell of
having to act on another object's internals in a separate external object.