Implement the MaxHeap data structure.
This max heap is based of the MaxHeap presented in section 5.5 of the textbook,
with some minor modifications.

The AbstractArrayTree has the methods used for getting the relative indexes of
nodes in a tree. I.e. it maps logical positions to physical positions.

The RecordMaxHeap uses these methods and more to ensure that it fulfils the heap
property.